+++ Modules/SharePointDsc/DSCResources/MSFT_SPUserProfileServiceApp/MSFT_SPUserProfileServiceApp.psm1
@@ -16,7 +16,7 @@ function Get-TargetResource
         [System.String]
         $ApplicationPool,
 
-        [Parameter()]
+        [Parameter(Mandatory = $true)]
         [System.String]
         $MySiteHostLocation,
 
@@ -72,12 +72,6 @@ function Get-TargetResource
     )
 
     Write-Verbose -Message "Getting user profile service application $Name"
-
-    if ($PSBoundParameters.ContainsKey("MySiteHostLocation") -eq $false)
-    {
-        Write-Verbose -Message ("You should also specify the MySiteHostLocation parameter. " + `
-                "This will be required as of SharePointDsc v4.0")
-    }
 
     $farmAccount = Invoke-SPDscCommand -Credential $InstallAccount `
         -Arguments $PSBoundParameters `
@@ -232,7 +226,7 @@ function Set-TargetResource
         [System.String]
         $ApplicationPool,
 
-        [Parameter()]
+        [Parameter(Mandatory = $true)]
         [System.String]
         $MySiteHostLocation,
 
@@ -288,12 +282,6 @@ function Set-TargetResource
     )
 
     Write-Verbose -Message "Setting user profile service application $Name"
-
-    if ($PSBoundParameters.ContainsKey("MySiteHostLocation") -eq $false)
-    {
-        Write-Verbose -Message ("You should also specify the MySiteHostLocation parameter. " + `
-                "This will be required as of SharePointDsc v4.0")
-    }
 
     if ($Ensure -eq "Present")
     {
@@ -526,7 +514,7 @@ function Test-TargetResource
         [System.String]
         $ApplicationPool,
 
-        [Parameter()]
+        [Parameter(Mandatory = $true)]
         [System.String]
         $MySiteHostLocation,
 
@@ -582,12 +570,6 @@ function Test-TargetResource
     )
 
     Write-Verbose -Message "Testing for user profile service application $Name"
-
-    if ($PSBoundParameters.ContainsKey("MySiteHostLocation") -eq $false)
-    {
-        Write-Verbose -Message ("You should also specify the MySiteHostLocation parameter. " + `
-                "This will be required as of SharePointDsc v4.0")
-    }
 
     $PSBoundParameters.Ensure = $Ensure